move platform view to app delegate.
