--- dev/integration_tests/wide_gamut_test/ios/Runner/DummyPlatformView.swift
@@ -1,68 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import Foundation
-
-import Flutter
-import UIKit
-
-class FLNativeViewFactory: NSObject, FlutterPlatformViewFactory {
-    private var messenger: FlutterBinaryMessenger
-
-    init(messenger: FlutterBinaryMessenger) {
-        self.messenger = messenger
-        super.init()
-    }
-
-    func create(
-        withFrame frame: CGRect,
-        viewIdentifier viewId: Int64,
-        arguments args: Any?
-    ) -> FlutterPlatformView {
-        return FLNativeView(
-            frame: frame,
-            viewIdentifier: viewId,
-            arguments: args,
-            binaryMessenger: messenger)
-    }
-
-    public func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
-          return FlutterStandardMessageCodec.sharedInstance()
-    }
-}
-
-class SolidColorView: UIView {
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-    setupView()
-  }
-
-  required init?(coder aDecoder: NSCoder) {
-    super.init(coder: aDecoder)
-    setupView()
-  }
-
-  private func setupView() {
-    backgroundColor = .blue
-  }
-}
-
-
-class FLNativeView: NSObject, FlutterPlatformView {
-    private var childView: UIView
-
-    init(
-        frame: CGRect,
-        viewIdentifier viewId: Int64,
-        arguments args: Any?,
-        binaryMessenger messenger: FlutterBinaryMessenger?
-    ) {
-        childView = SolidColorView()
-        super.init()
-    }
-
-    func view() -> UIView {
-        return childView
-    }
-}
